Adding Authorization

+++ app/graphql/mutations/delete_match_session.rb
@@ -5,6 +5,9 @@ class Mutations::DeleteMatchSession < Mutations::BaseMutation
   field :errors, [ String ], null: false
 
   def resolve(id:)
+    debugger
+    raise GraphQL::ExecutionError, "Login to access" unless context[:current_user]
+    raise GraphQL::ExecutionError, "Only admins can delete the blogs" unless context[:current_user].admin?
     match_session = MatchSession.find_by(id: id)
     if match_session
       match_session.destroy
